hough works!

+++ src/Bsuir.Misoi.Core/Images/Implementation/Hough/Line.cs
@@ -4,15 +4,15 @@ namespace Bsuir.Misoi.Core.Images.Implementation.Hough
 {
     public struct Line
     {
-        public Line(int f, int r)
+        public Line(double f, double r)
         {
             this.F = f;
             this.R = r;
             FRad = f * Math.PI / 180.0;
         }
 
-        public int F { get; }   //угол
-        public int R { get; }   //длина перпендикуляра
+        public double F { get; }   //угол
+        public double R { get; }   //длина перпендикуляра
         public double FRad { get; }
     }
 }
